Calculator v0.1.0 Calculator.Billage

Module to calculate the time and billing for the calls.

Functions

first(arg, acc \\ 0, minute \\ 0)

Applies the business rules to calls in the initial portion of time.

Based on the business rules, which implies that for an arbitrary initial period the billing is different, this function contains the rules for the initial period of each call. It is recursive on durations lower than the duration of the rule. And leaves after the minutes == factor is reached, leaving for the rules applied in Calculator.Billage.remainer/3.

For example, for a 5 minute rule with 300 seconds duration.

Examples

iex> Calculator.Billage.first({300, 5, 5})
{25, 0}
iex> Calculator.Billage.first({330, 5, 5})
{25, 30} # 30 seconds remaining

prepare(batch)

Batches the stream into an enumerable to be further used.

remainer(arg, factor, value)

Applies the business rules in the remainer time of the calls.

For the ending period of the call, this function applies its rules and return the total value, it must be called after Calculator.Billage.first/3.

Examples

# 1. Receives 5 minutes and a accumulated 25 cents with a factor of 1 which means it applies 2 cents per minute
# 2. Receives an extra 30 seconds

iex> Calculator.Billage.remainer({25, 300}, 1, 2) # 1.
35.0
iex> Calculator.Billage.remainer({25, 330}, 1, 2) # 2.
36.0