Calculator v0.1.0 Calculator.Sanitizer

Sanitizes the Record with errors and invalid parameters. This could be highly improved with the use of Ecto and changesets. But for this specific use case this module can be easily modified and works well.

Functions

sanitize(record)

Sanitizes a record and returns the validation status.

Examples

iex> record = %Calculator.Record{
...> call_from: "+351914374373",
...> call_to: "+351963433432",
...> end_time: "17:49:30",
...> start_time: "17:44:04"
...> }
iex> Calculator.Sanitizer.sanitize(record)
{:ok, %Calculator.Record{
  call_from: "+351914374373",
  call_to: "+351963433432",
  end_time: "17:49:30",
  start_time: "17:44:04"
}}

iex> invalid_time = %Calculator.Record{
...> call_from: "+351914374373",
...> call_to: "+351963433432",
...> end_time: "17:49:30",
...> start_time: :error # Coming from the parsing functions
...> }
...> Calculator.Sanitizer.sanitize(invalid_time)
{:error, %Calculator.Record{
  call_from: "+351914374373",
  call_to: "+351963433432",
  end_time: "17:49:30",
  start_time: :error
}, "INVALID TIME"}

iex> invalid_phone = %Calculator.Record{
...> call_from: "+35191437437a", # Mind the "a" at the end
...> call_to: "+35196343343", # Missing 1 number
...> end_time: "17:49:30",
...> start_time: :error # Coming from the parsing functions
...> }
...> Calculator.Sanitizer.sanitize(invalid_phone)
{:error, %Calculator.Record{
  call_from: "+35191437437a", # Mind the "a" at the end
  call_to: "+35196343343", # Missing 1 number
  end_time: "17:49:30",
  start_time: :error # Coming from the parsing functions
}, "INVALID PHONE"}