Calculator.Server (calculator v0.1.0)

Functions

calculate_fuel(mass, route)

Calculate fuel consumption for given mass & route.

Parameters

mass - Expressed as integer value in kilograms (kg).

route - List of tuples with direction & destination/gravity;

direction atom - either :launch or :land;

destination - bitstring;

One of "earth", "moon", "mars";

Capitalization does not matter, e.g. "Earth";

Atoms can be sent as well, e.g. :earth;

gravity - alternative to destination;

Should be a positive integer value;

Keyword list can be used for the route as well;

Examples

iex> Calculator.Server.calculate_fuel(
...>   28_801,
...>   launch: "earth",
...>   land: "moon",
...>   launch: "moon",
...>   land: "earth"
...> )
51_951

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_history()

Returns the calculator history as a list of entries in the following format:

- caller: #PID<...> of the caller;
- timestamp: UNIX timestamp in seconds;
- input: Map with the following values:

mass: positive integer value;

route: list of tuples with direction & destination/gravity;

output: positive integer value, result from the execution;

init(state)

Callback implementation for GenServer.init/1.

start_link(state \\ [])